+++ tests/test_response_chunked.py
@@ -1,0 +1,10 @@
+from bocadillo import API
+
+
+def test_chunked_response(api: API):
+    @api.route("/")
+    async def index(req, res):
+        res.chunked = True
+
+    r = api.client.get("/")
+    assert r.headers["transfer-encoding"] == "chunked"